requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
psycopg[binary]>=3.2.10
httpx==0.27.2
python-dotenv==1.0.1
